Clarify the ECN material in HOL-blocking
I made it clear that ECN is a beautiful thing where it is available, but it is not widely available, added some information about L4S while making clear that it is still being specified as an experimental mechanism, and tied this material to "QUIC Behavior" by adding discussion of QUIC HOL-blocking mitigation, in both RFC 9000 and https://datatracker.ietf.org/doc/html/draft-ietf-quic-datagram-02.

On balance, this removes my objection to the unqualified endorsement of ECN in the current editor's version, adds some indication that L4S might matter someday without endorsing it, and improves the QUIC behavior section.